Add proof-of-concept of server tabs
