+++ app/stats/@tabs/foo/page.tsx
@@ -1,0 +1,3 @@
+export default async function Page() {
+  return <div>This is foo.</div>;
+}
